HackPrinceton/FideGuide/backend/requirements.txt
Flask
openai
yfinance
python-dotenv
requests

